requirements.txt
transformers  
huggingface_hub  
streamlit 
evaluate  
pyvis  
python-dotenv  
git+https://github.com/huggingface/peft.git  
trl  
bitsandbytes
nltk
spacy
beautifulsoup4  
scikit-learn  